% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-toggle_subsections.R
\name{collapseMultilineCodeWorkhorse}
\alias{collapseMultilineCodeWorkhorse}
\title{Compress multiline code into single-line code}
\usage{
collapseMultilineCodeWorkhorse(codeToCompress)
}
\arguments{
\item{codeToCompress}{Code given as character string. The code must be without error!}
}
\value{
Character(1L) with the code being compressed
}
\description{
Compress multiline code into single-line code
}
\examples{
codeToCompress <- c(
  "f <- function(x) {",
  "  paste0(x, c('bla',",
  "  'bla'),",
  "  'bla'",
  "  )",
  "}",
  "f('asdf')"
)
collapseMultilineCodeWorkhorse(codeToCompress)
}
