% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-toggle_subsections.R
\name{copySection}
\alias{copySection}
\title{Copy section into clipboard}
\usage{
copySection()
}
\description{
Copy section into clipboard
}
\concept{importFrom rstudioapi getSourceEditorContext documentSave
importFrom clipr write_clip}
