% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-toggle_subsections.R
\name{deparse2}
\alias{deparse2}
\title{Featureful deparse:}
\usage{
deparse2(x, FLAGpad = TRUE)
}
\arguments{
\item{x}{an object to deparse}
}
\value{
deparsed code, collapsed into length 1
}
\description{
* data.frames are deparsed into tibble::tribble() calls with aligned columns
* vectors are dputted with one entry per row and aligned equal signs and commas
* Other objects are simply deparsed, with the shortest possible width.cutoff
}
\examples{
# Uncomment and try out
library(magrittr)
deparse2(1+1) \%>\% cat(sep = "\n")
deparse2(1:10 + 0.1) \%>\% cat(sep = "\n")
deparse2(setNames(1:26 + 0.1, letters)[1:10]) \%>\% cat(sep = "\n")
x <- c(wup = "sadfm", "sdflk" = "asldkfjlas", a = "asssdf", sldkfjlabs = "df")
deparse2(x) \%>\% cat(sep = "\n")
deparse2(list(
  a= 1:3,
  b = list("a", "b", "c"),
  d = list("a", "b", "c")
)) \%>\% cat(sep = "\n")
deparse2(data.frame(a = 1+1, b = c("c", "sdafksdfl"))) \%>\% cat(sep = "\n")
deparse2(function(x) {bla}) \%>\% cat(sep = "\n")
}
