% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-toggle_subsections.R
\name{duplicateArguments}
\alias{duplicateArguments}
\title{Turn fun(a, b) into fun(a = a, b = b)}
\usage{
duplicateArguments()
}
\value{

}
\description{
Turn fun(a, b) into fun(a = a, b = b)
}
\examples{
textline <- "wup <- fun(a, b)"
}
