% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-toggle_subsections.R
\name{exposeAsArgument}
\alias{exposeAsArgument}
\title{Like extract variable, but this writes it into the function argument section}
\usage{
exposeAsArgument()
}
\description{
Like extract variable, but this writes it into the function argument section
}
