% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-toggle_subsections.R
\name{extract_importFrom}
\alias{extract_importFrom}
\title{Go through code and extract all package::function calls into roxy tags}
\usage{
extract_importFrom()
}
\description{
Go through code and extract all package::function calls into roxy tags
}
\details{
rstudioapi getSourceEditorContext documentSave
}
