% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-toggle_subsections.R
\name{findFunctionCode}
\alias{findFunctionCode}
\title{Find rows which, taken together, are a function}
\usage{
findFunctionCode(documentText, row)
}
\arguments{
\item{documentText}{}

\item{row}{}
}
\value{

}
\description{
Find rows which, taken together, are a function
}
\examples{
documentText <- c(
  "fun1<- function(",
  "a,",
  "b = 1) {a^2}",
  "# blabla",
  "fun2 <- function(",
  "a,",
  "b = 1,",
  "d = c(a,",
  "b)) {
 a + b
 }",
  "1+2",
  "fun3 = function(",
  "a,",
  "b = 1) {a^2}",
  "# blabla"
)
row <- 12
lapply(1:length(documentText), function(i) findFunctionCode(documentText, i))
}
