% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-toggle_subsections.R
\name{findFunctionStartRow}
\alias{findFunctionStartRow}
\title{Title}
\usage{
findFunctionStartRow(documentText, row)
}
\arguments{
\item{documentText}{}

\item{row}{}
}
\description{
Title
}
