% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-toggle_subsections.R
\name{findRoxyIdxs}
\alias{findRoxyIdxs}
\title{Title}
\usage{
findRoxyIdxs(rowStart, documentText)
}
\arguments{
\item{rowStart}{start of function code, result from findFunctionStartRow()}

\item{documentText}{}
}
\value{
indices of the roxy skeleton
}
\description{
Title
}
