% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-toggle_subsections.R
\name{grepInUncommentedRange}
\alias{grepInUncommentedRange}
\title{In text, search for regex in the range of line:lineEnd and return the global index in text}
\usage{
grepInUncommentedRange(pattern = " *flplot *<- *", x, start, end)
}
\arguments{
\item{text}{}

\item{line}{}

\item{lineEnd}{}

\item{regex}{}
}
\value{
index as it would be returned by grep
}
\description{
In text, search for regex in the range of line:lineEnd and return the global index in text
}
\examples{
text <- c("pattern but out of range","asdf","pattern", "sadlkf")
grepInUncommentedRange(pattern = "pattern", x = text, start = 2, end = 3)
grepInUncommentedRange(pattern = "pattern", x = paste0("# ", text), start = 2, end = 3)
}
