% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-toggle_subsections.R
\name{insertDputNopad}
\alias{insertDputNopad}
\title{Execute selected text and insert its result as source code into your script}
\usage{
insertDputNopad()
}
\description{
data.frames are inserted as tibble::tribble
}
\examples{
# Uncomment and try out
1+1
x <- 1:10 + 0.1
x <- setNames(1:3 + 0.1,letters[1:3])
a <- list(
  a= 1:3,
  b = list("a", "b", "c"),
  d = list("a", "b", "c")
)
wup <- data.frame(a = 1+1, b ="c")
function(x) {bla}
}
