% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-toggle_subsections.R
\name{insertDputUnique}
\alias{insertDputUnique}
\title{insertDput, but wrap unique around the variable}
\usage{
insertDputUnique()
}
\description{
insertDput, but wrap unique around the variable
}
\examples{
# Uncomment and try out
1+1
x <- 1:10 + 0.1
x <- setNames(1:3 + 0.1,letters[1:3])
a <- list(
  a= 1:3,
  b = list("a", "b", "c"),
  d = list("a", "b", "c")
)
wup <- data.frame(a = 1+1, b ="c")
function(x) {bla}

# Or, basic texts
text <- "setNames(1:3 + 0.1,letters[1:3])"

}
