% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputting.R
\name{outputMdTable}
\alias{outputMdTable}
\title{Output table as markdown}
\usage{
outputMdTable(
  dt,
  split_by = NULL,
  filename = NULL,
  format = c("markdown", "pandoc"),
  caption = NULL,
  na.strings = ifelse(NFLAGtribble == 2, "NA", "-"),
  FLAGsummaryRow = TRUE,
  NFLAGtribble = 0,
  ...
)
}
\arguments{
\item{dt}{data.table}

\item{split_by}{Charactor of colnames to split the table by inserting separator lines}

\item{filename}{}
}
\value{
object of class knitr::kable
}
\description{
Output table as markdown
}
\examples{
tableName <- "bla"
dt <- data.table(iris)[1:5]
outputMdTable(dt)
outputMdTable(data.table(iris), split_by = "Species")
split_by <- NULL
filename <- NULL
format <- c("markdown", "pandoc")
caption <- NULL
na.strings <- "NA"
FLAGsummaryRow <- TRUE
NFLAGtribble <- 2
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
