% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-toggle_subsections.R
\name{projectPathComment_fromSection}
\alias{projectPathComment_fromSection}
\title{Add projectPath and projectComment automatically}
\usage{
projectPathComment_fromSection()
}
\value{

}
\description{
Transform 
# -------------------------------------------------------------------------#
# 15 estimate CL ----
# -------------------------------------------------------------------------#
}
\details{
to 
# -------------------------------------------------------------------------#
# 16 estimate CL ----
# -------------------------------------------------------------------------#
projectPath <- file.path(.modelFolder, "MODEL01") 
projectComment <- "estimate CL"
}
