% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-toggle_subsections.R
\name{renameFilesInPlaceAndInReport}
\alias{renameFilesInPlaceAndInReport}
\title{Work horse to rename files and update the file paths in all reports}
\usage{
renameFilesInPlaceAndInReport(
  ds,
  text,
  plotOrTab = c("plot", "tab"),
  .outputFolder
)
}
\arguments{
\item{ds}{}

\item{plotOrTab}{}
}
\value{

}
\description{
Work horse to rename files and update the file paths in all reports
}
