% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-toggle_subsections.R
\name{renumber_sections_filename}
\alias{renumber_sections_filename}
\title{Same as renumber_sections but with filename, not as addin}
\usage{
renumber_sections_filename(filename, FLAGfunctionAsSection = FALSE)
}
\description{
Same as renumber_sections but with filename, not as addin
}
