% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-toggle_subsections.R
\name{swapArg1Arg2}
\alias{swapArg1Arg2}
\title{Title}
\usage{
swapArg1Arg2()
}
\value{

}
\description{
e.g setdiff(x,y) becomes setdiff(y,x)
}
\examples{
# Uncomment and try
# setdiff(1:5, unique(c(1,2,6,7)))
}
