% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-toggle_subsections.R
\name{swapArg1Arg2InText}
\alias{swapArg1Arg2InText}
\title{Swap the first two arguments of a function call
e.g setdiff(x,y) becomes setdiff(y,x)}
\usage{
swapArg1Arg2InText(text)
}
\arguments{
\item{text}{}
}
\value{

}
\description{
Swap the first two arguments of a function call
e.g setdiff(x,y) becomes setdiff(y,x)
}
\examples{
text <- 'setdiff(1:5, unique(c(1,2,6,7)))'
swapArg1Arg2(text)
}
