% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-toggle_subsections.R
\name{toggle_blabla}
\alias{toggle_blabla}
\title{Toggle BLABLA so the documentwalker can extract functions}
\usage{
toggle_blabla()
}
\value{
NULL: Modifies document
}
\description{
Problem: the parser from Rstudio's extract function cannot handle code such as
foo <- bar[1,]
or 
foo <- bar[,1]
}
\details{
Thus wee first need to insert BLABLA into the specific places
foo <- bar[1,BLABLA]
or 
foo <- bar[BLABLA,1]

Then we can extract the function

rstudioapi getSourceEditorContext documentSave setDocumentContents
}
