% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-toggle_subsections.R
\name{toggle_roxyComments}
\alias{toggle_roxyComments}
\title{Switch between # and #' comments}
\usage{
toggle_roxyComments()
}
\description{
Switch between # and #' comments
}
\examples{
text <- c(
"# bla",
"# bla",
"  # bla",
"  # bla"
)

text <- c(
  "#' bla",
  "#' bla",
  "  #' bla",
  "  #' bla"
)
}
