% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-toggle_subsections.R
\name{toggleCommentForWholeSection}
\alias{toggleCommentForWholeSection}
\title{Comment/uncomment a whole section in one key stroke}
\usage{
toggleCommentForWholeSection()
}
\value{

}
\description{
Comment/uncomment a whole section in one key stroke
}
\concept{}
