% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-toggle_subsections.R
\name{turnIntoFactor}
\alias{turnIntoFactor}
\title{Title}
\usage{
turnIntoFactor()
}
\description{
Title
}
\examples{
dwup <- data.table::data.table(bla = c("wupwup", "bla"))
dwup[,`:=`(bla)]
}
