% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-toggle_subsections.R
\name{findConnectedCode}
\alias{findConnectedCode}
\title{Find rows which, taken together, are parseable code}
\usage{
findConnectedCode(documentText, row, FLAGreturnIdxs = FALSE)
}
\arguments{
\item{documentText}{}

\item{row}{}

\item{FLAGreturnIdxs}{return indices instead of text}
}
\description{
Find rows which, taken together, are parseable code
}
\examples{
documentText <- c(
  "a <- letters[",
  "1:3",
  "]",
  "b <- 1",
  "c(a,",
  "b)"
)
row <- 6
findConnectedCode(documentText, row)
for (i in 1:6) print(findConnectedCode(documentText, i))
}
